CC-35256: Fold ElasticsearchClientCloseTest's single test into ElasticsearchClientTest
awaitTerminationWithinSharesOneDeadlineAcrossPools tests a package-private
ElasticsearchClient helper, so it lives with the rest of the client tests;
one fewer single-test file.
